add uploadthing and implement image uploads

+++ src/app/api/candidates/[candidateId]/route.ts
@@ -4,6 +4,7 @@ import { z } from 'zod';
 import getCurrentUser from '@/app/actions/getCurrentUser';
 import prisma from '@/lib/prisma';
 import { CandidateValidator } from '@/lib/validators/candidate';
+import { utapi } from '@/app/api/uploadthing/core';
 
 interface IParams {
   params: {
@@ -27,18 +28,31 @@ export async function PUT(request: Request, { params }: IParams) {
       return new Response('ID no válido', { status: 400 });
     }
 
-    const body = CandidateValidator.parse(await request.json());
+    const { image, ...body } = CandidateValidator.parse(await request.json());
 
-    const candidate = await prisma.candidate.update({
+    const oldCandidate = await prisma.candidate.findUnique({
+      where: {
+        id: candidateId,
+      },
+    });
+
+    // if there is a new image, delete the old one
+    if (image && oldCandidate?.imageKey) {
+      await utapi.deleteFiles(oldCandidate?.imageKey);
+    }
+
+    const updatedCandidate = await prisma.candidate.update({
       where: {
         id: candidateId,
       },
       data: {
         ...body,
+        imageKey: image.key,
+        imageUrl: image.url,
       },
     });
 
-    return new Response(candidate.id);
+    return new Response(updatedCandidate.id);
   } catch (error) {
     if (error instanceof z.ZodError) {
       return new Response(error.message, { status: 422 });
@@ -66,7 +80,22 @@ export async function DELETE(request: Request, { params }: IParams) {
       return new Response('ID no válido', { status: 400 });
     }
 
-    const candidate = await prisma.candidate.delete({
+    const candidate = await prisma.candidate.findUnique({
+      where: {
+        id: candidateId,
+      },
+    });
+
+    if (!candidate) {
+      return new Response('No encontrado', { status: 404 });
+    }
+
+    // if there is a image, delete it
+    if (candidate?.imageKey) {
+      await utapi.deleteFiles(candidate.imageKey);
+    }
+
+    await prisma.candidate.delete({
       where: {
         id: candidateId,
       },
